backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
azure-ai-textanalytics==5.1.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.24.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
frozenlist==1.3.0
idna==3.3
isodate==0.6.1
msrest==0.7.1
multidict==6.0.2
oauthlib==3.2.0
PyMySQL==1.0.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
yarl==1.7.2
